Update Details
userDetail, videoDetail
and protect path

+++ controllers/videoController.js
@@ -39,8 +39,11 @@ export const postUpload = async (req, res) => {
   const newVideo = await Video.create({
     fileUrl: path,
     title,
-    description
+    description,
+    creator: req.user.id
   });
+  req.user.videos.push(newVideo.id);
+  req.user.save();
   res.redirect(routes.videoDetail(newVideo.id));
 };
 
@@ -49,7 +52,7 @@ export const videoDetail = async (req, res) => {
     params: { id }
   } = req;
   try {
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate('creator');
     res.render('videoDetail', { pageTitle: video.title, video });
   } catch (error) {
     console.log(error);
@@ -63,6 +66,9 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
+    if (video.creator !== req.user.id) {
+      throw Error();
+    }
     res.render('editVideo', { pageTitle: `Edit ${video.title}`, video });
   } catch (error) {
     console.log(error);
@@ -89,6 +95,10 @@ export const deleteVideo = async (req, res) => {
     params: { id }
   } = req;
   try {
+    const video = await Video.findById(id);
+    if (video.creator !== req.user.id) {
+      throw Error();
+    }
     await Video.findByIdAndRemove({ _id: id });
   } catch (error) {
     console.log(error);
